add simple integ test against existing collection
We have now finalized our collection data format and will begin giving some customers early access to fqlx.  This test is going to ensure that we don't make any backwards incompatible changes to our collection data format.  The plan is to run this after the dev and preview deploys in the database pipeline.

We also plan to run the full drivers test suite in these pipelines, to ensure we don't have any backwards incompatible drivers changes.  Some of those tests don't currently work when pointed at dev, taking an incremental approach here and will be getting this test in the pipeline first while I work on resolving the other drivers tests.  At that point will switch to running all of the integ tests in the pipeline.