add schema version to drivers
ENG-5237
This builds on a change to core that started including the schema version in the query API response:
https://github.com/fauna/core/blob/master/ext/api/src/main/resources/openapi.yml#L393
This schema version can be used by the dashboard to determine when it needs to refresh the schema.  This includes both displaying schema to users as well as refreshing the environment used by the language library to provide intellisense.